Excluindo comandos inúteis
