+++ commands/fun/boanoite.js
@@ -2,7 +2,7 @@ const Discord = require('discord.js');
 
 module.exports = {
   run: async (client, message, args) => {
-    message.reply('Bom noite! 🌙')
+    message.reply('Boa noite! 🌙')
   },
 
   conf: {},
