+++ commands/fun/tempo.js
@@ -28,13 +28,13 @@ module.exports = {
       const tempoinfo = new Discord.MessageEmbed()
       .setAuthor(`☁️ Previsão do tempo para ${current.observationpoint}`)
       .setThumbnail(current.imageUrl)
-      .addField('Fuso horário', `UTC${location.timezone}`, true)
-      .addField('Tipo de grau', 'Celsius', true)
-      .addField('Temperatura', `${current.temperature}°`, true)
-      .addField('Vento', current.winddisplay, true)
-      .addField('Sensação', `${current.feelslike}°`, true)
-      .addField('Humidade', `${current.humidity}%`, true)
-      .setColor("#29C9FC")
+      .addField('⌚ Fuso horário', `UTC${location.timezone}`, true)
+      .addField('📍 Tipo de grau', 'Celsius', true)
+      .addField('🌡️ Temperatura', `${current.temperature}°`, true)
+      .addField('🌪 Vento', current.winddisplay, true)
+      .addField('☁️ Sensação', `${current.feelslike}°`, true)
+      .addField('💧 Humidade', `${current.humidity}%`, true)
+      .setColor(process.env.COLOR)
       .setFooter('2021 © Liga dos Programadores.')
       .setTimestamp()
       message.channel.send(tempoinfo)
