+++ commands/info/avatar.js
@@ -1,5 +1,5 @@
 /**
- * O Comando "avatar" mostrará a imagem de perfil do usuário ou do bot
+ * O Comando "avatar" mostrará a imagem de perfil do usuário.
 */
 
 const Discord = require("discord.js");
@@ -10,15 +10,15 @@ module.exports = {
   * Que passará os argumentos atraves do middleware.
 */
 
-    run: function (client, message, _args) {
+    run: function (client, message, args) {
         const member = message.mentions.users.first() || message.author;
 
-	const embed = new Discord.MessageEmbed()
-    .setTitle(`Avatar de ${member.username}`)
-    .setImage(member.displayAvatarURL({ format: 'png', dynamic: true, size: 1024 }))
-    .setColor(process.env.COLOR)
-    .setFooter('2021 © Liga dos Programadores.')
-    .setTimestamp()
+        const embed = new Discord.MessageEmbed()
+        .setAuthor(`🖼️ Avatar de ${member.username}`)
+        .setImage(member.displayAvatarURL({ format: 'png', dynamic: true, size: 1024 }))
+        .setColor(process.env.COLOR)
+        .setFooter('2021 © Liga dos Programadores.')
+        .setTimestamp()
 
         message.channel.send(embed);
     },
@@ -26,14 +26,14 @@ module.exports = {
     conf: {},
 
     /**
-   * Aqui exportamos ajuda do comando como o seu nome categoria, descrição, etc...
-  */
+        * Aqui exportamos ajuda do comando como o seu nome categoria, descrição, etc...
+    */
 
     get help() {
         return {
             name: "avatar",
-            category: "Info",
-            description: "Mostra o avatar do usuário ou de um bot.",
+            category: "Informações",
+            description: "Mostra o avatar do usuário.",
             usage: "avatar"
         };
     }
