+++ commands/info/serverinfo.js
@@ -35,7 +35,7 @@ module.exports = {
       .addField('**Canais de voz**', channels.filter(channel => channel.type === 'voice').size, true)
       .addField('**Criado em**', formatDate('DD/MM/YYYY, às HH:mm:ss', date), true)
       .addField('**Você entrou em**', formatDate('DD/MM/YYYY, às HH:mm:ss', joined), true)
-      .addField('**Cargos**', `[${roles.lenght -1}]`, roles.lenght <10 ? roles.join(`, `) : roles.lenght >10 ? this.client.utils.trimArray(roles) : 'Nenhum cargo')
+      .addField('**Cargo(s)**', `${roles.length < 10 ? roles.join(', ') : roles.length > 10 ? this.client.utils.trimArray(roles) : 'Nenhum'}`)
       .setFooter('2021 © Liga dos Programadores.')
       .setTimestamp()
 
@@ -53,7 +53,7 @@ module.exports = {
   get help () {
     return {
       name: 'serverinfo',
-      category: 'Info',
+      category: 'Informações',
       description: 'Informação sobre o servidor',
       usage: 'serverinfo'
     }
