+++ events/apresentar.js
@@ -3,26 +3,21 @@
  */
 
 module.exports = async (client, message) => {
-  const presentedRole = process.env.ID_ROLE_APRESENTACAO;
-  const presentationChannel = message.guild.channels.cache.get(
-    process.env.APRESENTACAO
-  );
-
+  const presentedRole = process.env.CARGO_APRESENTOU;
+  const presentationChannel = message.guild.channels.cache.find((channel) => channel.id == process.env.CANAL_APRESENTACAO);
+  
   if (presentationChannel && !message.member.roles.cache.has(presentedRole)) {
     message.member.roles.add(presentedRole).catch(console.error);
-
-    const emoji = message.guild.emojis.cache.get(
-      process.env.APRESENTACAO_EMOJI_ID
-    );
+    const emoji = message.guild.emojis.cache.get(process.env.SERVER_EMOJI);
     emoji && message.react(emoji);
     console.log(message.member)
     message.member.roles.add(presentedRole).catch(console.error)
   } else {
     const embed = {
-      color: 16739451,
-      title: 'Como resetar seu status de apresentação:',
+      color: process.env.COLOR,
+      title: 'Como resetar seu status de apresentação',
       description:
-        '**Hey**, caso você tenha errado a digitação de algo em sua mensagem de apresentação, basta digitar o comando `p!reset` no *chat de comandos do servidor* para resetar a sua apresentação!',
+        'Caso você tenha errado a digitação de algo em sua mensagem de apresentação, basta digitar o comando `!resetar` no *chat de comandos do servidor* para resetar a sua apresentação!',
     };
     message.author
       .send({ embed: embed })
