+++ events/boasvindas.js
@@ -1,0 +1,8 @@
+const Discord = require('discord.js');
+require('dotenv').config();
+
+module.exports = (client, message, args) => {
+  let guild = client.guilds.cache.get(" ");
+  let channel = client.channels.cache.get(" ");
+
+}
