+++ events/message.js
@@ -9,7 +9,7 @@ module.exports = async (client, message) => {
    */
 
   if (message.author.bot) return;
-  if (message.channel.id == process.env.APRESENTACAO)
+  if (message.channel.id == process.env.CANAL_APRESENTACAO)
     return client.emit('apresentar', message);
 
   /** Outra boa pratica é ignorar qualquer mensagem que não começe com o prefixo escolhido do bot.
